Поправил внешний вид информации о вкладе. Исправил тип данных баланса на Double - это логично, т.к. после начисления процентов баланс будет дробным.

+++ app/src/main/java/com/example/sandboxbank/App/core/deposit/data/db/FinancialEntity.kt
@@ -13,6 +13,6 @@ data class FinancialEntity(
     val percentRate: Double,
     val percentType: Int,
     val period: Int,
-    val balance: Long,
+    val balance: Double,
     val name: String
 )
